+++ apps/io-session-manager-internal/src/__mocks__/services/session-service.mock.ts
@@ -1,6 +1,7 @@
 import { vi } from "vitest";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import { SessionService } from "../../services/session-service";
+import { anUnlockedUserSessionState } from "../user.mock";
 
 export const mockInvalidateUserSession = vi
   .fn()
@@ -20,10 +21,15 @@ export const mockUnlockUserAuthentication = vi
 
 export const mockDeleteUserSession = vi.fn().mockReturnValue(RTE.right(null));
 
+export const mockGetUserSessionState = vi
+  .fn()
+  .mockReturnValue(RTE.right(anUnlockedUserSessionState));
+
 export const SessionServiceMock: SessionService = {
   invalidateUserSession: mockInvalidateUserSession,
   getUserSession: mockGetUserSession,
   lockUserAuthentication: mockLockUserAuthentication,
   unlockUserAuthentication: mockUnlockUserAuthentication,
   deleteUserSession: mockDeleteUserSession,
+  getUserSessionState: mockGetUserSessionState,
 };